move core_init to `vault.rs` & generate Nostr ID by `bs58` instead of `libp2p`
