+++ relay/Cargo.toml
@@ -11,5 +11,5 @@ openssl = "0.10.63"
 hex = "0.4.3"
 lazy_static = "1.4.0"
 secp256k1 = { version = "0.27.0", features = ["serde", "bitcoin-hashes", "bitcoin-hashes-std", "rand"] }
-libp2p = { version = "0.53.0", features = ["secp256k1"] }
+bs58 = "0.5.1"
 serde = { version = "1.0.117", features = ["derive"] }
