+++ relay/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod pki;
+pub mod vault;
